refactor #128: Registration, MyAccount 페이지 리팩토링
- UserProfile 컴포넌트 구현
- Logout, LocationSelectField 구현
