+++ fe/src/components/MyAccount/Logout/style.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const AddLocationButtonBox = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: flex-end;
+  height: 30vh;
+  width: 100vw;
+`;
