+++ fe/src/components/Registration/LocationSelectField/style.ts
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const AddLocationButtonBox = styled.div`
+  display: flex;
+  justify-content: center;
+  width: 100vw;
+`;
+
+export const LocationBox = styled.div`
+  height: 5vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
