+++ fe/src/components/common/UserProfile/style.ts
@@ -24,26 +24,8 @@ export const NoticeBox = styled.div`
   align-items: center;
 `;
 
-export const LocationBox = styled.div`
-  height: 5vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const Notice = styled.div`
-  padding-left: 10px;
-  ${({ theme }) => theme.font.fontType.footNote};
-`;
-
 export const UserId = styled.div`
   padding-left: 5px;
   ${({ theme }) => theme.font.fontType.title2};
   color: ${({ theme }) => theme.color.colors.neutralTextStrong};
-`;
-
-export const AddLocationButtonBox = styled.div`
-  display: flex;
-  justify-content: center;
-  width: 100vw;
 `;
