+++ fe/src/pages/MyAccount/index.tsx
@@ -1,13 +1,21 @@
-import UserInfo from '@Components/MyAccount/UserInfo';
+import Logout from '@Components/MyAccount/Logout';
 import { NavigationBar } from '@Components/common/NavBar';
 import { TabBarHome } from '@Components/common/TabBar';
+import UserProfile from '@Components/common/UserProfile';
+import { useUserInfoContext } from '@Contexts/userInfoContext';
 
-const MyAccount = () => (
-  <>
-    <NavigationBar type="defaultLayout" title="내 계정" />
-    <UserInfo />
-    <TabBarHome currentPage="account" />
-  </>
-);
+const MyAccount = () => {
+  const { userInfo } = useUserInfoContext();
+
+  return (
+    <>
+      <NavigationBar type="defaultLayout" title="내 계정" />
+      <UserProfile avatar={userInfo.avatar} username={userInfo.username}>
+        <Logout />
+      </UserProfile>
+      <TabBarHome currentPage="account" />
+    </>
+  );
+};
 
 export default MyAccount;
